refactor: move ./config to ./src/config
