+++ apps/server/src/controllers/temp.controller.js
@@ -1,12 +1,12 @@
-import { status } from '../../config/response.status';
-import { getTempData } from '../services/temp.service';
-import { response } from '../../config/response';
+import { status } from "../config/response.status";
+import { getTempData } from "../services/temp.service";
+import { response } from "../config/response";
 
 export const tempTest = (req, res, next) => {
   res.send(response(status.SUCCESS, getTempData()));
 };
 
 export const tempException = (req, res, next) => {
-  console.log('temp/exception/' + req.params.flag);
+  console.log("temp/exception/" + req.params.flag);
   return res.send(response(status.SUCCESS, CheckFlag(req.params.flag)));
 };
