+++ apps/server/src/services/github.service.js
@@ -1,7 +1,7 @@
-import { BaseError } from '../../config/error';
-import { status } from '../../config/response.status';
-import { githubClient } from '../github/client';
-import { contributionQuery } from '../github/query';
+import { BaseError } from "../config/error";
+import { status } from "../config/response.status";
+import { githubClient } from "../github/client";
+import { contributionQuery } from "../github/query";
 
 export const getUserContributions = async (username) => {
   try {
